requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==21.2.0
azure-core==1.16.0
azure-functions==1.7.2
azure-functions-durable==1.0.3
azure-storage-blob==12.8.1
certifi==2021.5.30
cffi==1.14.6
chardet==4.0.0
charset-normalizer==2.0.3
cryptography==3.4.7
furl==2.1.2
idna==3.2
iniconfig==1.1.1
isodate==0.6.0
msrest==0.6.21
multidict==5.1.0
oauthlib==3.1.1
orderedmultidict==1.0.1
packaging==21.0
pluggy==0.13.1
py==1.10.0
pycparser==2.20
pyparsing==2.4.7
pytest==6.2.4
python-dateutil==2.8.2
requests==2.26.0
requests-oauthlib==1.3.0
six==1.16.0
toml==0.10.2
typing-extensions==3.10.0.0
urllib3==1.26.6
yarl==1.6.3
